故障告警记录修改

+++ src/main/java/com/xiliulou/electricity/vo/failureAlarm/FailureWarnFrequencyVo.java
@@ -50,5 +50,5 @@ public class FailureWarnFrequencyVo {
     /**
      * 使用天数
      */
-    private Integer usageDays;
+    private Long usageDays;
 }
